TextboxGenerator: begin work on better error reporting
Inline TextboxRenderer.render into TextboxGenerator.doInBackground
Remove TextboxRenderer (no longer needed)
