+++ src/main/java/adudecalledleo/aftbg/app/ui/worker/TextboxGenerator.java
@@ -1,17 +1,25 @@
 package adudecalledleo.aftbg.app.ui.worker;
 
+import java.awt.*;
+import java.awt.image.*;
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.swing.*;
 
-import adudecalledleo.aftbg.app.TextboxRenderer;
 import adudecalledleo.aftbg.app.data.DataTracker;
 import adudecalledleo.aftbg.app.data.Textbox;
+import adudecalledleo.aftbg.app.face.Face;
 import adudecalledleo.aftbg.app.game.GameDefinition;
+import adudecalledleo.aftbg.app.text.DOMParser;
+import adudecalledleo.aftbg.app.text.DOMRenderer;
+import adudecalledleo.aftbg.app.text.node.Node;
 import adudecalledleo.aftbg.app.text.node.color.ColorParser;
 import adudecalledleo.aftbg.app.ui.LoadFrame;
 import adudecalledleo.aftbg.app.ui.MainPanel;
 import adudecalledleo.aftbg.app.ui.dialog.PreviewDialog;
+import adudecalledleo.aftbg.app.util.ColorUtils;
+import adudecalledleo.aftbg.app.util.Pair;
 import adudecalledleo.aftbg.window.WindowContext;
 
 public final class TextboxGenerator extends AbstractWorker {
@@ -30,12 +38,43 @@ public final class TextboxGenerator extends AbstractWorker {
 
     @Override
     protected Void doInBackground() {
-        // FIXME TextboxRenderer.render just hides errors!
-        //  Inline it, then report errors properly here.
+        final int textboxCount = textboxes.size();
+        var image = new BufferedImage(816, 180 * textboxCount + 2 * (textboxCount - 1), BufferedImage.TYPE_INT_ARGB);
 
-        var image = TextboxRenderer.render(winCtx, parserCtx, textboxes);
+        var g = image.createGraphics();
+        g.setBackground(ColorUtils.TRANSPARENT);
+        g.clearRect(0, 0, image.getWidth(), image.getHeight());
 
-        if (image == null) {
+        List<Pair<Integer, DOMParser.Error>> errors = null;
+        for (int i = 0; i < textboxCount; i++) {
+            var textbox = textboxes.get(i);
+            var result = DOMParser.parse(textbox.getText(), parserCtx);
+            if (result.hasErrors()) {
+                if (errors == null) {
+                    errors = new ArrayList<>();
+                }
+                for (var error : result.errors()) {
+                    errors.add(new Pair<>(i, error));
+                }
+                continue;
+            }
+
+            // only render textbox if we don't have any errors (since if we do have errors,
+            //  we're just gonna throw away the result anyway)
+            if (errors == null) {
+                g.setClip(0, 182 * i, 816, 180);
+                renderTextbox(winCtx, g, 0, 182 * i, textbox.getFace(), result.document(), i < textboxCount - 1 ? 0 : -1);
+            }
+        }
+
+        g.dispose();
+
+        if (errors == null) {
+            var dialog = new PreviewDialog(mainPanel, image);
+            dialog.setLocationRelativeTo(null);
+            cleanup();
+            dialog.setVisible(true);
+        } else {
             loadFrame.setAlwaysOnTop(false);
             // TODO more detailed error message
             JOptionPane.showMessageDialog(mainPanel,
@@ -43,14 +82,20 @@ public final class TextboxGenerator extends AbstractWorker {
                             + "Correct this, then try generating again.",
                     "Generate textbox(es)", JOptionPane.ERROR_MESSAGE);
             cleanup();
-        } else {
-            var dialog = new PreviewDialog(mainPanel, image);
-            dialog.setLocationRelativeTo(null);
-            cleanup();
-            dialog.setVisible(true);
         }
 
         return null;
     }
 
+    private static void renderTextbox(WindowContext winCtx, Graphics2D g, int x, int y, Face face, Node root, int arrowFrame) {
+        winCtx.drawBackground(g, x + 4, y + 4, 808, 172, null);
+        g.drawImage(face.getImage(), x + 18, y + 18, null);
+        DOMRenderer.render(g, root,
+                x + (face.isBlank() ? 18 : 186),
+                y + 21);
+        winCtx.drawBorder(g, x, y, 816, 180, null);
+        if (arrowFrame >= 0) {
+            winCtx.drawArrow(g, x, y, 816, 180, arrowFrame, null);
+        }
+    }
 }
